Moved favorite-related files into a single directory
