+++ react-app/src/components/Favorite-Components/FavoritesPage/index.js
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { Redirect } from "react-router-dom";
 import * as favoriteActions from "../../store/favorite";
 import FavoritesTile from "../FavoritesTile";
-import NoFavorites from "./NoFavorites";
+import NoFavorites from "../NoFavorites";
 import "./FavoritesPage.css";
