+++ react-app/src/components/Favorite-Components/FavoritesTile/index.js
